% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_refactor.R
\name{cm_refactor}
\alias{cm_refactor}
\title{Helper function to make confusion matrices easier}
\usage{
cm_refactor(predvar, refvar)
}
\arguments{
\item{predvar}{Predictor variable}

\item{refvar}{Reference variable}
}
\value{
A caret confusion matrix
}
\description{
Refactor two inputs and make confusion matrix with caret package
}
\author{
Joseph Brehm
}
