% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVIRCurve.R
\name{CVIRSigmoid}
\alias{CVIRSigmoid}
\alias{CVIRComputeSCurve}
\alias{CVIRTrapezoid}
\alias{CVIRBeta}
\alias{CVIRGauss}
\alias{CVIRTriangle}
\alias{CVIRPI}
\alias{CVIRLinear}
\title{CVIR Evaluation Curves}
\usage{
CVIRSigmoid(x, xlim, ascending = TRUE)

CVIRComputeSCurve(x, xlim, ascending = TRUE)

CVIRTrapezoid(x, xlim)

CVIRBeta(x, xlim)

CVIRGauss(x, xlim)

CVIRTriangle(x, xlim)

CVIRPI(x, xlim)

CVIRLinear(x, xlim = NULL)
}
\arguments{
\item{x}{a vector of property values}

\item{xlim}{x-axis limits (see details)}

\item{ascending}{should S-curve be drawn in ascending or descending order? Default: \code{TRUE}}
}
\value{
a vector of fuzzy rating values derived from specified curve equation
}
\description{
CVIR Evaluation Curves
}
\details{
describe the number of xlim parameters needed for each curve type here
}
\examples{

x <- seq(0, 4, 0.01)

y <- CVIRSigmoid(x, c(0.5, 3))
plot(y ~ x)


x <- seq(0, 10, 0.01)

y <- CVIRTrapezoid(x, 2:5)
plot(y ~ x)


x <- seq(0, 10, 0.01)

y <- CVIRBeta(x, c(2,2))
plot(y ~ x)


x <- seq(0, 10, 0.01)

y <- CVIRGauss(x, c(2,2))
plot(y ~ x)


x <- seq(0, 10, 0.01)

y <- CVIRTriangle(x, c(1,3))
plot(y ~ x)


x <- seq(0, 10, 0.01)

y <- CVIRPI(x, c(4, 1))
plot(y ~ x)

x <- seq(0, 10, 0.01) 
y <- CVIRLinear(x, c(4, 1))
plot(y ~ x)
}
