% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{datatree_hsg}
\alias{datatree_hsg}
\title{Cached Hydrologic Soil Group data.tree object}
\format{
An object of class \code{Node} (inherits from \code{R6}) of length 45.
}
\source{
NASIS Calculation
}
\usage{
datatree_hsg
}
\description{
Cached Hydrologic Soil Group data.tree object
}
\keyword{datasets}
