% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{datatree_svi}
\alias{datatree_svi}
\title{Cached Soil Vulnerability Index data.tree object}
\format{
An object of class \code{Node} (inherits from \code{R6}) of length 46.
}
\source{
NASIS Interpretation
}
\usage{
datatree_svi
}
\description{
Cached Soil Vulnerability Index data.tree object
}
\keyword{datasets}
