% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_eval.R
\name{dwb_calc}
\alias{dwb_calc}
\title{Dwewllings Without Basements}
\usage{
dwb_calc(indata)
}
\arguments{
\item{indata}{Input data (data.frame or SpatRaster)}
}
\value{
Evaluation result
}
\description{
Dwewllings Without Basements
}
