% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{evalbyeid}
\alias{evalbyeid}
\title{Evaluation by the eid code}
\usage{
evalbyeid(eid, d, sig.scale = 1)
}
\arguments{
\item{eid}{Evaluation ID -- NB: not rule id or property id. Evaluation only.}

\item{d}{data to pass through evaluation curve?}

\item{sig.scale}{default 1}
}
\value{
result of evaluation made with the specified evaluation curve
}
\description{
Evaluation by the eid code
}
