% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{extractArbitraryCurveEval}
\alias{extractArbitraryCurveEval}
\title{Extract Arbitrary Eval Curve}
\usage{
extractArbitraryCurveEval(x, invert)
}
\arguments{
\item{x}{evaluation curve XML text}

\item{invert}{logical}
}
\value{
curve function
}
\description{
Extract Arbitrary Eval Curve
}
