% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractArbitraryLinearCurveEval}
\alias{extractArbitraryLinearCurveEval}
\title{Extract Arbitrary Linear Eval Curve}
\usage{
extractArbitraryLinearCurveEval(x, invert)
}
\arguments{
\item{x}{evaluation curve XML text}

\item{invert}{logical}
}
\value{
curve function
}
\description{
Extract Arbitrary Linear Eval Curve
}
