% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{extractCrispCurveEval}
\alias{extractCrispCurveEval}
\title{Extract Crisp Evaluation Curve}
\usage{
extractCrispCurveEval(x, invert, res, dmin, dmax)
}
\arguments{
\item{x}{evalulation curve XML text}

\item{invert}{logical}

\item{res}{number of intermediate points}

\item{dmin}{domain range minimum}

\item{dmax}{domain range maximum}
}
\value{
curve function
}
\description{
Extract Crisp Evaluation Curve
}
