% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractCrispExpression}
\alias{extractCrispExpression}
\title{Extract Crisp Expression Logic as R function}
\usage{
extractCrispExpression(x, invert = FALSE, asString = FALSE)
}
\arguments{
\item{x}{\emph{character}. Evaluation XML content containing a CrispExpression}

\item{invert}{\emph{logical}. Invert logic with \code{!}? Default: \code{FALSE}}

\item{asString}{\emph{logical}. Return un-parsed function (for
debugging/inspection) Default: \code{FALSE}}
}
\value{
\emph{function}. A generated function of an input variable \code{x}
}
\description{
Extract Crisp Expression Logic as R function
}
\details{
The generated function returns a logical value (converted to
numeric) when the relevant property data are supplied.
}
