% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractEvalCurve}
\alias{extractEvalCurve}
\title{Extract an Evaluation Curve}
\usage{
extractEvalCurve(
  evalrec,
  xlim = NULL,
  as.list = FALSE,
  resolution = NULL,
  sig.scale = NULL
)
}
\arguments{
\item{evalrec}{data.frame. A single evaluation record from \code{NASIS_evaluations} data set.}

\item{xlim}{numeric. Length 2. Minimum and maximum values to override internal domain range.}

\item{as.list}{Return list with function and domain limits? Default \code{FALSE} returns just the
function.}

\item{resolution}{Deprecated. Not used.}

\item{sig.scale}{Deprecated. Not used.}
}
\value{
function that can evaluate input data and return evaluation values. When \code{as.list=TRUE} a
list object containing elements \code{FUN} (the function) and \code{domain} (a numeric vector containing
domain minimum and maximum)
}
\description{
Extract an Evaluation Curve
}
