% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractIsNull}
\alias{extractIsNull}
\title{Extract IsNull Evaluation Logic as R function}
\usage{
extractIsNull(invert = FALSE)
}
\arguments{
\item{invert}{invert logic with \code{!}? Default: \code{FALSE}}
}
\value{
a generated function of an input variable \code{x}
}
\description{
Default behavior of IsNull evaluation returns value > \code{0} (1) if \code{NULL}, inverted behavior returns \code{0} if \code{NULL}
}
\details{
The generated function returns a logical value (converted to numeric) when the relevant property data are supplied.
}
