% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{extractLinearCurveEval}
\alias{extractLinearCurveEval}
\title{Extract Linear Evaluation Curve}
\usage{
extractLinearCurveEval(x, invert, res)
}
\arguments{
\item{x}{evaluation curve XML text}

\item{invert}{logical}

\item{res}{number of intermediate points}
}
\value{
curve function
}
\description{
Extract Linear Evaluation Curve
}
