% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractSigmoidCurveEval}
\alias{extractSigmoidCurveEval}
\title{Extract Sigmoid Eval Curve}
\usage{
extractSigmoidCurveEval(x, xlim, invert)
}
\arguments{
\item{x}{evaluation curve XML text}

\item{xlim}{domain range minimum/maximum}

\item{invert}{logical}
}
\value{
curve function
}
\description{
Extract Sigmoid Eval Curve
}
