% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractTrapezoidEval}
\alias{extractTrapezoidEval}
\title{Extract Trapezoidal Eval Curve}
\usage{
extractTrapezoidEval(x, xlim, invert = FALSE)
}
\arguments{
\item{x}{evaluation curve XML text}

\item{invert}{logical}
}
\value{
curve function
}
\description{
Extract Trapezoidal Eval Curve
}
