% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{getAttributeByEval}
\alias{getAttributeByEval}
\title{Get Attribute By Evaluation}
\usage{
getAttributeByEval(x, a)
}
\arguments{
\item{x}{a data.tree}

\item{a}{attribute}
}
\value{
attribute
}
\description{
Get Attribute By Evaluation
}
