% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{getPropertySet}
\alias{getPropertySet}
\title{Get unique properties for all evaluations}
\usage{
getPropertySet(x)
}
\arguments{
\item{x}{vector of evaluations}
}
\value{
a set of properties
}
\description{
Get unique properties for all evaluations
}
