% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{initRuleset}
\alias{initRuleset}
\title{Initialize a ruleset}
\usage{
initRuleset(rulename)
}
\arguments{
\item{rulename}{Rule name character}
}
\value{
ruleset
}
\description{
Initialize a ruleset
}
