% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret}
\alias{interpret}
\alias{interpret,Node,data.frame-method}
\alias{interpret,character,data.frame-method}
\alias{interpret,Node,SpatRaster-method}
\alias{interpret,character,SpatRaster-method}
\title{Run a Rule on Custom Property Data}
\usage{
interpret(x, propdata, ...)

\S4method{interpret}{Node,data.frame}(x, propdata, mode = "node", cache = FALSE, ...)

\S4method{interpret}{character,data.frame}(x, propdata, mode = "node", cache = FALSE, ...)

\S4method{interpret}{Node,SpatRaster}(
  x,
  propdata,
  mode = "node",
  cores = 1,
  core_thresh = 250000L,
  file = paste0(tempfile(), ".tif"),
  nrows = nrow(propdata)/(terra::ncell(propdata)/core_thresh),
  overwrite = TRUE,
  ...
)

\S4method{interpret}{character,SpatRaster}(
  x,
  propdata,
  mode = "node",
  cores = 1,
  core_thresh = 250000L,
  file = paste0(tempfile(), ".tif"),
  nrows = nrow(propdata)/(terra::ncell(propdata)/core_thresh),
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{x}{A \emph{data.tree} object containing rule tree, or a \emph{character} giving the rule name to load with  \code{\link[=initRuleset]{initRuleset()}}.}

\item{propdata}{\emph{data.frame} or \emph{SpatRaster} object with column names corresponding to input properties. The column names should be named using \code{make.names(propname)} where \code{propname} is the property name from \link{NASIS_properties} data object.}

\item{...}{Additional arguments}

\item{mode}{character. Either \code{"table"} or \code{"node"} (default) . Controls back-end rating calculation method using base R or data.tree, respectively.}

\item{cache}{logical. Save input property data in data.tree object? Default: \code{FALSE}}

\item{cores}{integer. Default \code{1} core.}

\item{core_thresh}{integer. Default \code{250000} cells.}

\item{file}{character. Path to output raster file. Defaults to a temporary GeoTIFF.}

\item{nrows}{integer. Default \code{nrow(propdata) / (terra::ncell(propdata) / core_thresh)}}

\item{overwrite}{logical. Overwrite \code{file} if it exists?}
}
\value{
\emph{data.frame} containing \code{"rating"}. When \code{mode="node"} the input object \code{x} is modified in place as a side effect with \code{"rating"} values. When \code{cache=TRUE} the input \code{"data"} values are also stored within each node.
}
\description{
Allows for evaluation of a primary rule including subrules, operators, hedges, and evaluations.
}
\details{
The user must supply a \emph{data.frame} object or SpatRaster object with property data as input.
}
\examples{

r <- initRuleset("Erodibility Factor Maximum")
p <- getPropertySet(r)

my_data <- data.frame(Kmax = seq(0, 1, 0.01))
colnames(my_data) <- make.names(p$propname)

res <- interpret(r, my_data)

plot(res$rating ~ my_data[[1]],
     xlab = "K factor (input)",
     ylab = "Rating [0-1]")


}
