% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretationEngine-package.R
\docType{package}
\name{InterpretationEngine-package}
\alias{InterpretationEngine}
\alias{InterpretationEngine-package}
\title{InterpretationEngine: Stand-alone NASIS Interpretation Engine}
\description{
InterpretationEngine contains tools to develop, test, and apply interpretations to soil data outside the context of NASIS. This project aims to create a prototype from existing interpretation rules, evaluations, and properties as managed in NASIS. Once the prototype is complete it should be possible to generate fuzzy ratings from arbitrary sources of soil and environmental data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://ncss-tech.github.io/interpretation-engine/}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

Authors:
\itemize{
  \item Dylan Beaudette \email{dylan.beaudette@usda.gov}
  \item Joseph Brehm \email{jrbrehm@nmsu.edu}
}

}
\keyword{internal}
