% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{linkEvaluationFunctions}
\alias{linkEvaluationFunctions}
\title{Link Evaluation Functions}
\usage{
linkEvaluationFunctions(node)
}
\arguments{
\item{node}{a data.tree node}
}
\value{
a modified data.tree object
}
\description{
Link Evaluation Functions
}
