% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{linkHedgeOperatorFunctions}
\alias{linkHedgeOperatorFunctions}
\title{Link Hedge and Operator Functions}
\usage{
linkHedgeOperatorFunctions(node)
}
\arguments{
\item{node}{a data.tree node}
}
\value{
a modified data.tree object
}
\description{
Link Hedge and Operator Functions
}
