% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{linkSubRules}
\alias{linkSubRules}
\title{Link Subrules}
\usage{
linkSubRules(node)
}
\arguments{
\item{node}{a data.tree node}
}
\value{
a modified data.tree object
}
\description{
Link Subrules
}
