% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{lookupProperties}
\alias{lookupProperties}
\title{Lookup Properties using NASIS Web Report}
\usage{
lookupProperties(coiid, propIDs)
}
\arguments{
\item{coiid}{Vector of component IDs (\code{coiid})}

\item{propIDs}{Vector of property IDs}
}
\value{
properties?
}
\description{
This function uses \code{WEB-PROPERY-COMPONENT_property} NASIS Web Report to lookup property data
}
