% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupRatingClass.R
\name{lookupRatingClass}
\alias{lookupRatingClass}
\title{Lookup Rating Classes by Rule ID}
\usage{
lookupRatingClass(ruleiidref)
}
\arguments{
\item{ruleiidref}{Rule ID}
}
\value{
A data.frame containing columns \code{ruleiidref} (Rule ID), \code{ratingclassname}, \code{ratingclassupperboundary} and \code{ruleratingclassiid}
}
\description{
Lookup Rating Classes by Rule ID
}
