% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{makeNamesUnique}
\alias{makeNamesUnique}
\title{serial number added to names}
\usage{
makeNamesUnique(l)
}
\arguments{
\item{l}{a data.tree}
}
\value{
a data.tree
}
\description{
serial number added to names
}
