% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{makeNamesUnique2}
\alias{makeNamesUnique2}
\title{Use hash function for unique names}
\usage{
makeNamesUnique2(l)
}
\arguments{
\item{l}{a data.tree}
}
\value{
a data.tree
}
\description{
Use hash function for unique names
}
