% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{makeNamesUnique3}
\alias{makeNamesUnique3}
\title{lookup rule name}
\usage{
makeNamesUnique3(l)
}
\arguments{
\item{l}{a data.tree}
}
\value{
a data.tree
}
\description{
Split Rule ref Ids from Evaluation ref Ids
}
