% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{NASIS_evaluations}
\alias{NASIS_evaluations}
\title{Cached NASIS Evaluations}
\format{
An object of class \code{data.frame} with 10858 rows and 13 columns.
}
\source{
NASIS 7.x evaluations
}
\usage{
NASIS_evaluations
}
\description{
Cached NASIS Evaluations
}
\keyword{datasets}
