% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{NASIS_properties}
\alias{NASIS_properties}
\title{Cached NASIS Properties}
\format{
An object of class \code{data.frame} with 8096 rows and 8 columns.
}
\source{
NASIS 7.x properties
}
\usage{
NASIS_properties
}
\description{
Cached NASIS Properties
}
\keyword{datasets}
