% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{NASIS_property_def}
\alias{NASIS_property_def}
\title{Cached NASIS Property Definitions}
\format{
An object of class \code{data.frame} with 7837 rows and 3 columns.
}
\source{
NASIS 7.x property definitions
}
\usage{
NASIS_property_def
}
\description{
Cached NASIS Property Definitions
}
\keyword{datasets}
