% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{NASIS_rules}
\alias{NASIS_rules}
\title{Cached NASIS Rules}
\format{
An object of class \code{data.frame} with 13502 rows and 7 columns.
}
\source{
NASIS 7.x rules
}
\usage{
NASIS_rules
}
\description{
Cached NASIS Rules
}
\keyword{datasets}
