% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{parseRule}
\alias{parseRule}
\title{Convert interpretation rule into data.tree representation}
\usage{
parseRule(x)
}
\arguments{
\item{x}{should contain \code{rule} element with XML text to parse}
}
\value{
a data.tree
}
\description{
Convert interpretation rule into data.tree representation
}
