% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propByPropname.R
\name{propByPropname}
\alias{propByPropname}
\title{Property by Property Name}
\usage{
propByPropname(propname)
}
\arguments{
\item{propname}{Character. Property Name}
}
\value{
a row from the \code{NASIS_properties} data.frame
}
\description{
Property by Property Name
}
\examples{
propByPropname("SOIL REACTION 1-1 WATER IN DEPTH 0-100cm (min)")
}
