% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propByPropname.R
\name{propdefByPropname}
\alias{propdefByPropname}
\title{Property Definition by Property Name}
\usage{
propdefByPropname(propname)
}
\arguments{
\item{propname}{Character. Property Name}
}
\value{
a row from the \code{NASIS_property_def} data.frame
}
\description{
Property Definition by Property Name
}
\examples{
propdefByPropname("SOIL REACTION 1-1 WATER IN DEPTH 0-100cm (min)")
}
