% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_SDA.R
\name{pull_SDA}
\alias{pull_SDA}
\title{Quick SDA data acquisition
This function is a shortcut for loading in a broad set of SDA data, filtered to area symbols specified by an input character vector.}
\usage{
pull_SDA(asym, fun = "in")
}
\arguments{
\item{asym}{character. Area symbol}

\item{fun}{Default: "in" uses SQL "IN" operator. Alternately: "like".}
}
\value{
SDA table result
}
\description{
The following tables are returned:
}
\details{
\itemize{
\item \code{legend}
\item \code{mapunit}
\item \code{component}
\item \code{chorizon}
\item \code{corestrictions}
\item \code{cosurffrags}
\item \code{cotaxfmmin}
\item \code{codiagfeatures}
\item \code{comonth}
\item \code{chtexturegrp}
\item \code{chfrags}
\item \code{chtexture}
\item \code{cosoilmoist}
\item \code{muaggatt}
\item \code{chunified}
\item \code{cointerpENG} -- "ENG" rules only though
}
}
\author{
Joseph Brehm
}
