% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_SDA_compandup.R
\name{pull_SDA_compup}
\alias{pull_SDA_compup}
\title{Quick SDA data acquisition (component up)}
\usage{
pull_SDA_compup(asym, fun = "in")
}
\arguments{
\item{asym}{character. Area symbol}

\item{fun}{Default: "in" uses SQL "IN" operator. Alternately: "like".}
}
\value{
list object containing SDA table information for legend, mapunit, component
}
\description{
This function is a shortcut for loading in a broad set of SDA data, filtered to area symbols specified by an input character vector.
}
\details{
The following tables are returned:
\itemize{
\item \code{legend}
\item \code{mapunit}
\item \code{component}
}
}
\author{
Joseph Brehm
}
