% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleByRuleName.R
\name{ruleByRulename}
\alias{ruleByRulename}
\title{Rule by Rule Name}
\usage{
ruleByRulename(rulename)
}
\arguments{
\item{rulename}{Character. Rule Name}
}
\value{
a row from the \code{NASIS_rules} data.frame
}
\description{
Rule by Rule Name
}
\examples{
ruleByRulename("Erodibility Factor Maximum")
}
