% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{tree_eval}
\alias{tree_eval}
\alias{hsg_calc}
\alias{svi_calc}
\title{Evaluate a tree}
\usage{
tree_eval(tree, indata, ncores = 1)
}
\arguments{
\item{tree}{tree is a data tree object structured with classifications as leaves above those, all nodes must contain these attributes: "var", specifying the variable name to evaluate, "logical", specifying the logical evaluation at that step. MUST INCLUDE THE SAME VARIABLE "VAR"}

\item{indata}{data must be a data frame containing columns with names matching all the values of "var" used in the tree object}

\item{ncores}{number of cores to parallelize over. if set to 1 (default), runs sequentially. Also accepts "auto" as input. Will run with ncores on computer - 2.}
}
\value{
eval result TODO
}
\description{
Evaluate a tree
}
\author{
Joseph Brehm
}
