% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_eval.R
\name{vf_calc}
\alias{vf_calc}
\title{Valley Fever}
\usage{
vf_calc(indata, doxeric = "auto")
}
\arguments{
\item{indata}{Input data (as data.frame or SpatRaster)}

\item{doxeric}{force xeric or nonxeric function, or let it choose for each data row/pixel depending on an input raster (accepts "auto" / "xeric" / "nonxeric")}
}
\value{
Evaluation result (as data.frame or SpatRaster)
}
\description{
Valley Fever
}
\author{
Joseph Brehm
}
