% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{xmlChunkParse}
\alias{xmlChunkParse}
\title{Parse an evaluation chunk XML return a list}
\usage{
xmlChunkParse(x)
}
\arguments{
\item{x}{evaluation curve XML text}
}
\value{
parsed XML chunk
}
\description{
Parse an evaluation chunk XML return a list
}
