% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_refactor.R
\name{cm_refactor}
\alias{cm_refactor}
\title{Helper function to make confusion matricies easier
Refactor two inputs and make confusion matrix with caret package}
\usage{
cm_refactor(predvar, refvar)
}
\arguments{
\item{predvar}{predictor var}

\item{refvar}{reference var}
}
\value{
a caret confusion matrix
}
\description{
Helper function to make confusion matricies easier
Refactor two inputs and make confusion matrix with caret package
}
\author{
Joseph Brehm
}
