% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{dwb_calc}
\alias{dwb_calc}
\title{Dwewllings Without Basements}
\usage{
dwb_calc(indata)
}
\arguments{
\item{indata}{Input data}
}
\value{
evaluation result
}
\description{
Dwewllings Without Basements
}
