% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{evalByEID}
\alias{evalByEID}
\alias{evalbyeid}
\title{Extract Evaluation Curve by Evaluation ID}
\usage{
evalByEID(eid, d, sig.scale = 1)

evalbyeid(eid, d, sig.scale = 1)
}
\arguments{
\item{eid}{Evaluation ID -- NB: not rule id or property id. Evaluation only.}

\item{d}{data to pass through evaluation curve?}

\item{sig.scale}{not used}
}
\value{
result of evaluation made with the specified evaluation curve
}
\description{
Extract Evaluation Curve by Evaluation ID
}
