% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractEvalCurve}
\alias{extractEvalCurve}
\title{Extract an evaluation curve}
\usage{
extractEvalCurve(evalrec, xlim = NULL, resolution = NULL, sig.scale = NULL)
}
\arguments{
\item{evalrec}{Evaluation record}

\item{xlim}{numeric vector, min and max values of fuzzy space}

\item{resolution}{not used}

\item{sig.scale}{not used}
}
\value{
evaluation curve values
}
\description{
Extract an evaluation curve
}
