% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractEvalCurve}
\alias{extractEvalCurve}
\title{Extract an Evaluation Function}
\usage{
extractEvalCurve(evalrec, xlim = NULL, as.list = FALSE)
}
\arguments{
\item{evalrec}{\emph{data.frame}. A single evaluation record from
\code{NASIS_evaluations} data set.}

\item{xlim}{\emph{numeric}. Length 2. Minimum and maximum values to override
internal domain range.}

\item{as.list}{\emph{logical}. Return  a \emph{list} with function and domain limits?
Default \code{FALSE} returns just the function.}
}
\value{
An R function that takes arbitrary input data and returns evaluation
values. When \code{as.list=TRUE} a list object containing elements \code{FUN} (the
function) and \code{domain} (a numeric vector containing domain minimum and
maximum)
}
\description{
Extract an R function from a NASIS evaluation for to evaluate curves and
expressions.
}
