% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationCurves.R
\name{extractEvalCurve}
\alias{extractEvalCurve}
\title{Extract an evaluaton curve}
\usage{
extractEvalCurve(evalrec, resolution = 250, sig.scale = 1)
}
\arguments{
\item{evalrec}{Evaluation record}

\item{resolution}{Number of intermediate points}

\item{sig.scale}{default 1}
}
\value{
evaluation curve values
}
\description{
Extract an evaluaton curve
}
