% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{extractSigmoidCurveEval}
\alias{extractSigmoidCurveEval}
\title{Extract Sigmoid Eval Curve}
\usage{
extractSigmoidCurveEval(x, invert, res, sig.scale = 1)
}
\arguments{
\item{x}{evaluation curve XML text}

\item{invert}{logical}

\item{res}{number of intermediate points}

\item{sig.scale}{default 1}
}
\value{
curve function
}
\description{
Extract Sigmoid Eval Curve
}
