% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{extractTrapezoidEval}
\alias{extractTrapezoidEval}
\title{Extract Trapeziodal Eval Curve}
\usage{
extractTrapezoidEval(x, invert)
}
\arguments{
\item{x}{evaluation curve XML text}

\item{invert}{logical}
}
\value{
curve function
}
\description{
Extract Trapeziodal Eval Curve
}
