% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-functions.R
\name{getAndCacheData}
\alias{getAndCacheData}
\title{Cache dataset containing important NASIS data
soilDB::uncode() used to convert coded -> uncoded values}
\usage{
getAndCacheData()
}
\value{
cached data
}
\description{
Cache dataset containing important NASIS data
soilDB::uncode() used to convert coded -> uncoded values
}
