% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret}
\alias{interpret}
\alias{interpret,Node,data.frame-method}
\alias{interpret,character,data.frame-method}
\alias{interpret,Node,SpatRaster-method}
\alias{interpret,character,SpatRaster-method}
\title{Run a Rule on Custom Property Data}
\usage{
interpret(x, propdata, ...)

\S4method{interpret}{Node,data.frame}(x, propdata, cache = FALSE, ...)

\S4method{interpret}{character,data.frame}(x, propdata, cache = FALSE, ...)

\S4method{interpret}{Node,SpatRaster}(
  x,
  propdata,
  cores = 1,
  core_thresh = 250000L,
  file = paste0(tempfile(), ".tif"),
  nrows = nrow(propdata)/(terra::ncell(propdata)/core_thresh),
  overwrite = TRUE,
  ...
)

\S4method{interpret}{character,SpatRaster}(
  x,
  propdata,
  cores = 1,
  core_thresh = 250000L,
  file = paste0(tempfile(), ".tif"),
  nrows = nrow(propdata)/(terra::ncell(propdata)/core_thresh),
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{x}{A \emph{data.tree} Object containing Rule tree (e.g. from \code{\link[=initRuleset]{initRuleset()}})}

\item{propdata}{A data.frame or SpatRaster object}

\item{...}{Additional arguments}

\item{cache}{logical. Save input property data in data.tree object? Default: \code{FALSE}}

\item{cores}{integer. Default \code{1} core.}

\item{core_thresh}{integer. Default \code{25000} cells.}

\item{file}{character. Path to output raster file. Defaults to a temporary GeoTIFF.}

\item{nrows}{integer. Default \code{nrow(propdata) / (terra::ncell(propdata) / core_thresh)}}

\item{overwrite}{logical. Overwrite \code{file} if it exists?}
}
\value{
\emph{data.frame} containing \code{"rating"}
}
\description{
Allows for evaluation of a primary rule including subrules, operators, hedges, and evaluations.
}
\details{
The user must supply a \emph{data.frame} object or SpatRaster object with property data as input.
The column names should be named using \code{make.names(propname)} where \code{propname} is the property name from
\link{NASIS_properties} data object in this package.

\code{cache} argument stores input data in x along with evaluation ratings
}
\examples{

r <- initRuleset("Erodibility Factor Maximum")
p <- getPropertySet(r)

my_data <- data.frame(Kmax = seq(0, 1, 0.01))
colnames(my_data) <- make.names(p$propname)

res <- interpret(r, my_data)

plot(res$rating ~ my_data[[1]],
     xlab = "K factor (input)",
     ylab = "Rating [0-1]")


}
