% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupProperties.R
\name{lookupProperties}
\alias{lookupProperties}
\title{Lookup Properties using NASIS Web Report}
\usage{
lookupProperties(coiid, propIDs)
}
\arguments{
\item{coiid}{Vector of component IDs (\code{coiid})}

\item{propIDs}{Vector of property IDs}
}
\value{
A data.frame containing \code{propiid}, \code{coiid}, \code{comp_name}, \code{comp_pct} and the representative value for the property (\code{rv}).
}
\description{
This function uses \code{WEB-PROPERY-COMPONENT_property} NASIS Web Report to look up component property data
}
