% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvaluation.R
\name{plotEvaluation}
\alias{plotEvaluation}
\title{Plot Evaluation Curve}
\usage{
plotEvaluation(x, xlim = NULL, resolution = 100, pch = NA, ...)
}
\arguments{
\item{x}{data}

\item{xlim}{\code{plot} xlim default \code{NULL}}

\item{resolution}{number of points (default 100)}

\item{pch}{point character for source domain/range values. Default: \code{NA}}

\item{...}{additional arguments passed to plot}
}
\value{
a plot
}
\description{
Plot Evaluation Curve
}
