% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_SDA2.r
\name{pull_SDA}
\alias{pull_SDA}
\title{Quick SDA data acquisition
This function is a shortcut for loading in a broad set of SDA data, filtered to area symbols specified by an input character vector.}
\usage{
pull_SDA(asym, fun = "in")
}
\arguments{
\item{asym}{area symbol}

\item{fun}{default "in"; alternately: "like}
}
\value{
SDA table result
}
\description{
Other than filtering by area symbol, no modifications are made to the data
at last edit, these tables are returned:
}
\details{
\itemize{
\item \code{legend}
\item \code{mapunit}
\item \code{component}
\item \code{chorizon}
\item \code{corestrictions}
\item \code{cosurffrags}
\item \code{cotaxfmmin}
\item \code{codiagfeatures}
\item \code{comonth}
\item \code{chtexturegrp}
\item \code{chfrags}
\item \code{chtexture}
\item \code{cosoilmoist}
\item \code{muaggatt}
\item \code{chunified}
\item \code{cointerpENG} -- "ENG" rules only though
}
}
\author{
Joseph Brehm
}
