% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_SDA_compandup.R
\name{pull_SDA_compup}
\alias{pull_SDA_compup}
\title{Quick SDA data acquisition (component up)
This function is a shortcut for loading in a broad set of SDA data, filtered to area symbols specified by an input character vector.}
\usage{
pull_SDA_compup(asym, fun = "in")
}
\arguments{
\item{asym}{area symbol}

\item{fun}{default "in"; alternately: "like}
}
\value{
SDA table result
}
\description{
Other than filtering by area symbol, no modifications are made to the data
at last edit, these tables are returned:
}
\details{
\itemize{
\item \code{legend}
\item \code{mapunit}
\item \code{component}
}
}
\author{
Joseph Brehm
}
